fix: address review findings from PR #7
- restart_instance_logic and stop_instance_logic now call
  with_self_host_network_extravars so self-host instances receive
  firewall_mode='helper' and qlsm_network_rules on restart/stop,
  matching the pattern already used by deploy/start/apply_config/delete

- Move resolve_self_host_management_target and its helpers
  (detect_docker_host_ip, _can_resolve_hostname,
  _detect_gateway_from_proc_route) from ui/routes/self_host_helpers
  into ui/task_logic/self_host_network — task logic should not import
  from routes; update server_status_poll and standalone_host_setup
  imports and fix test_self_host_helpers to reference the new location

- Add security notice to docs/setup.md documenting that host-init
  runs with privileged:true + pid:host
