+++ tests/test_self_host_helpers.py
@@ -5,6 +5,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from ui.routes import self_host_helpers as helpers
+from ui.task_logic import self_host_network as net
 
 
 def test_detect_docker_host_ip_from_proc_route(tmp_path):
@@ -14,7 +15,7 @@ def test_detect_docker_host_ip_from_proc_route(tmp_path):
         "eth0\t00000000\t010011AC\t0003\t0\t0\t0\t00000000\n"
     )
 
-    assert helpers.detect_docker_host_ip(route_path=route_file) == "172.17.0.1"
+    assert net.detect_docker_host_ip(route_path=route_file) == "172.17.0.1"
 
 
 def test_detect_docker_host_ip_falls_back_to_ip_route(monkeypatch, tmp_path):
@@ -24,9 +25,9 @@ def test_detect_docker_host_ip_falls_back_to_ip_route(monkeypatch, tmp_path):
         assert cmd == ["ip", "route", "show", "default"]
         return subprocess.CompletedProcess(cmd, 0, stdout="default via 172.18.0.1 dev eth0\n", stderr="")
 
-    monkeypatch.setattr(helpers.subprocess, "run", fake_run)
+    monkeypatch.setattr(net.subprocess, "run", fake_run)
 
-    assert helpers.detect_docker_host_ip(route_path=route_file) == "172.18.0.1"
+    assert net.detect_docker_host_ip(route_path=route_file) == "172.18.0.1"
 
 
 def test_detect_docker_host_ip_raises_when_unavailable(monkeypatch, tmp_path):
@@ -36,44 +37,44 @@ def test_detect_docker_host_ip_raises_when_unavailable(monkeypatch, tmp_path):
     def fake_run(cmd, check, capture_output, text):
         return subprocess.CompletedProcess(cmd, 1, stdout="", stderr="no route")
 
-    monkeypatch.setattr(helpers.subprocess, "run", fake_run)
+    monkeypatch.setattr(net.subprocess, "run", fake_run)
 
     with pytest.raises(ValueError, match="Could not detect host machine IP"):
-        helpers.detect_docker_host_ip(route_path=route_file)
+        net.detect_docker_host_ip(route_path=route_file)
 
 
 def test_resolve_self_host_management_target_prefers_env_override(monkeypatch):
     monkeypatch.setenv("QLSM_SELF_HOST_SSH_TARGET", "10.10.10.10")
 
-    assert helpers.resolve_self_host_management_target() == "10.10.10.10"
+    assert net.resolve_self_host_management_target() == "10.10.10.10"
 
 
 def test_resolve_self_host_management_target_prefers_host_docker_internal(monkeypatch):
     monkeypatch.delenv("QLSM_SELF_HOST_SSH_TARGET", raising=False)
     monkeypatch.setattr(
-        helpers,
+        net,
         "_can_resolve_hostname",
         lambda name: name == "host.docker.internal",
     )
     monkeypatch.setattr(
-        helpers,
+        net,
         "detect_docker_host_ip",
         lambda **_: "172.18.0.1",
     )
 
-    assert helpers.resolve_self_host_management_target() == "host.docker.internal"
+    assert net.resolve_self_host_management_target() == "host.docker.internal"
 
 
 def test_resolve_self_host_management_target_falls_back_to_gateway(monkeypatch):
     monkeypatch.delenv("QLSM_SELF_HOST_SSH_TARGET", raising=False)
-    monkeypatch.setattr(helpers, "_can_resolve_hostname", lambda name: False)
+    monkeypatch.setattr(net, "_can_resolve_hostname", lambda name: False)
     monkeypatch.setattr(
-        helpers,
+        net,
         "detect_docker_host_ip",
         lambda **_: "172.18.0.1",
     )
 
-    assert helpers.resolve_self_host_management_target() == "172.18.0.1"
+    assert net.resolve_self_host_management_target() == "172.18.0.1"
 
 
 def test_detect_default_self_ssh_user_prefers_env(monkeypatch, tmp_path):
